Restructured the repo to allow multiple issuers
This will be used for testing with an invalid issuer in platform/data-service.

The top-level directory now contains a directory for each issuer.
`issuer-0` is the pre-existing issuer. `issuer-1` is an alternative
issuer for testing against.

In order to not break the platform build (until corresponding changes
have been merged there), issuer-0 is duplicated at its previous location
of ~/.well-known. The duplication will be removed in a subsequent PR.